Move ImmutableList<T>.Node to a new file
